(#112) refactor: activity 스키마 파일명 및 경로 일괄 변경
