+++ src/domain/Activity/components/main/CategorySelect.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { GetActivitiesRequestQuery } from '@/domain/Activity/schemas/main/activity';
+import { GetActivitiesRequestQuery } from '@/domain/Activity/schemas/main';
 import { cn } from '@/shared/libs/cn';
 
 const categories: NonNullable<GetActivitiesRequestQuery['category']>[] = [
